Add new controller method: VehicleEditDetails
Implemented a new method, VehicleEditDetails, in the VehiclesController class. This method handles the data retrieving of a user vehicle needed for editing.

The VehicleEditDetails method accepts the necessary order details as parameters and performs the following steps:

Validates the input data to ensure it meets the required criteria - does the vehicle exist and is the user its owner.
Retrieves the necessary information from the database.