Add new controller method: EditVehicle
Implemented a new method, EditVehicle, in the VehiclesController class. This method handles request for vehicle details editing by the user.

The EditVehicle method accepts the necessary details as parameters and performs the following steps:

Validates the input data to ensure it meets the required criteria - does the vehicle exist and is the user its owner.
Validates that the model state of the parameters is valid.