Add new controller method: Edit
Implemented a new method, EditVehicle, in the VehiclesController class. This method handles request for vehicle details editing by the user.

The Edit method accepts the necessary details as parameters and performs the following steps:

Validates the input data to ensure it meets the required criteria - does the tripexist and is the user its creator.
Validates that the model state of the parameters is valid.
If the validation steps are valid - the service method for trip editing is called.