Add: new component EditTrip
Created a new functional component, EditTrip, to display the trip details on the UI. This component is responsible for rendering the trip record data retrieved from the server and displaying it in a structured format.
The user can edit the different fields.
When the editing process is completed and all the fields are valid a request is send to save the changes in the server.