Add new controller method: Delete
Implemented a new method, Delete, in the TripController class. This method handles request for trip record deleting by the user.

The Deletemethod accepts the trip record identifier.

Validates the input data to ensure it meets the required criteria - does the trip exist and is the user its creator.
If the validation steps are valid - the service method for trip record deleting is called.