Add: new component ServiceRecords
Created a new functional component, ServiceRecords, to display the trip overview details on the UI. This component is responsible for rendering the service records retrieved from the server and displaying it in a structured format.
The retrieved data will be passed to a card component.
The user will have an option to delete or edit a service record.