Add: new  controller method - ServiceRecordDetails
Implemented a new method, ServiceRecordDetails, in the ServiceRecordController. This method handles request for retrieving a specific user service record.

The ServiceRecordDetails method accepts the necessary details as parameters and performs the following steps:

Validates the input data to ensure it meets the required criteria - does the service record exist and is the user its creator.
If the validation steps are valid - the service method for retrieving the service record is called.