Add: new controller method - Edit
Implemented a new patch method, Edit, in the ServiceRecordsController class. This method handles request for editing service records details by the user.

The Edit method accepts the necessary details as parameters and performs the following steps:

Validates the input data to ensure it meets the required criteria - does the record exists and is the user its creator.
Validates that the model state of the parameters is valid.
If the validation steps are valid - the service method for service record editing is called.