Add: new component - TaxRecords
Created a new functional component, TaxRecords, to display the trip overview details on the UI. This component is responsible for rendering the tax records retrieved from the server and displaying it in a structured format.
Additionally a field with tax statistics will be available for the user.
The retrieved data will be passed to a card component.
The user will have an option to delete or edit a tax record.