Add: new controller method - Delete
Implemented a new method, Delete, in the TaxRecordsController class. This method handles request for deleting the tax record.

The TaxRecordsController accepts the tax record id.

Validates the input data to ensure it meets the required criteria:
- Does the user exist
- Does the tax record exist
- Is the user the tax record owner

If the validation steps are valid - the service method for tax record delete is called.