Add: new controller methods
Implemented two  new method, TaxRecordsCount and TaxRecordsCost, in the TaxRecordsController class.

1. The TaxRecordsCount :
Validates the input data to ensure it meets the required criteria:
- Does the user exist.
If the validation steps are valid - the service method for retrieving the tax records count is called.

2. The TaxRecordsCost:
Validates the input data to ensure it meets the required criteria:
- Does the user exist.
If the validation steps are valid - the service method for retrieving the tax records cost is called.