Add: new controller methods
Implemented two new method, ServiceRecordsCount and ServiceRecordsCost, in the ServiceRecordsController class.

1. The ServiceRecordsCount :
Validates the input data to ensure it meets the required criteria:
- Does the user exist.
If the validation steps are valid - the service method for retrieving the service records count is called.

2. The ServiceRecordsCost:
Validates the input data to ensure it meets the required criteria:
- Does the user exist.
If the validation steps are valid - the service method for retrieving the service records cost is called.