Add: new controller method - LastNCountServiceRecordsByUserId
Implemented a new method, LastNCountServiceRecordsByUserId, in the ServiceRecordsController class. This method handles request for retrieving information about an n-count service records.
The LastNCountServiceRecordsByUserId accepts the count of records to be retrieved.

Validates the input data to ensure it meets the required criteria:
- Does the user exist

If the validation steps are valid - the service method for service record retrieving is called.