Major update: controller methods
Modified the controller to implement working with refresh tokens.

Edit:
1. Register - the register no longer generates a JWT token for direct sign in on the user.
2. Login - the login generates a JWT token and a HttpOnly Secure Cookie containing the generated Refresh token

Add: new controller method - RefreshUserToken
Implemented a new method, RefreshUserToken, in the IdentityControllerclass. This method handles request for refreshing the user JWT token or canceling a Refresh token to make the user log in again.
The RefreshUserToken accepts the count a JWT token.

Validates the following criterias:
- Does an JWT token exists
- Does a Refresh token exist
- Are the JWT token principals valid
- Is the user the refresh token owner
- Is the token expired

If the validation steps are valid - the service method for refreshing the JWT token is called.