Add: controller method Logout
Implemented a new method, Logout, in the IdentityController class. This method handles request for logging out the user.
The Logout method retrieves the user id from the JWT token claims.

Validates the input data to ensure it meets the required criteria:
- Does the user exist

If the validation steps are valid - the service method for terminating the Refresh token is called.